+++ src/mtezq.c
@@ -307,10 +307,10 @@ error_t mt_mergeQueue(MTQueue* new_queue, Queue** queues, size_t num_queues)
 {
 	if(new_queue != NULL)
 	{
-		for(int i = 0; i < num_queues; i++)
+		for(size_t i = 0; i < num_queues; i++)
 		{
 			size_t q_len = queues[i]->length;
-			for(int j = 0; j < q_len; j++)
+			for(size_t j = 0; j < q_len; j++)
 			{
 				mt_enqueue(new_queue, dequeue(queues[i]));
 			}
@@ -329,10 +329,10 @@ error_t mt_mergeMTQueue(MTQueue* new_queue, MTQueue** queues, size_t num_queues)
 {
 	if(new_queue != NULL)
 	{
-		for(int i = 0; i < num_queues; i++)
+		for(size_t i = 0; i < num_queues; i++)
 		{
 			size_t q_len = queues[i]->length;
-			for(int j = 0; j < q_len; j++)
+			for(size_t j = 0; j < q_len; j++)
 			{
 				mt_enqueue(new_queue, mt_dequeue(queues[i]));
 			}
